hope ur happy now build, also updated to latest exiled

+++ Lone079/Config.cs
@@ -12,7 +12,7 @@ namespace Lone079
 
 		public int HealthPercent { get; set; } = 50;
 		public string BroadcastMessage { get; set; } = "<i>You have been respawned as a random SCP with half health because all other SCPs have died.</i>";
-		public int BroadcastTime { get; set; } = 10;
+		public ushort BroadcastTime { get; set; } = 10;
 
 		public List<RoleType> scp079Respawns { get; set; } = new List<RoleType>
 		{
